- Issue 108: Allocine Plugin New MovieDatabasePlugin using a french movie database site named AlloCine.
- Issue 109: Multiple Id for Movie objects
Change the current id to a map of id indexed by a pluginIdString.
For example :
   mediaFile.setId(ImdbPlugin.IMDB_PLUGIN_ID, imdbId);
   mediaFile.setId(AllocinePlugin.ALLOCINE_PLUGIN_ID, imdbId);

   mediaFile.getId(ImdbPlugin.IMDB_PLUGIN_ID);

also
   mediaFile.getId();
   mediaFile.setId(imdbId);
will be kept for compatibility purpose.

- Issue 78: Custom coverart Enhancement
Quick first implementation (not flexible but working ;) )
Scan the .nfo file for the URL of poster (currently only URL starting with http:// and ending with .jpg or .JPG)
If this URL exists, set it as poster URL and to not search for another poster.