Add to Issu 101 : add attribute "moviedb" to <id> tag for other plugins
In the XBMC xml nfo file we now support (see 
http://code.google.com/p/moviejukebox/issues/detail?id=101 ) there is only
one <id> tag allowed which can be used to specify IMDB Id.
I can be usefull to use several different Id for different movieDb working
with movieDbPlugin.

We now support the addition of an attribute "moviedb" to the <id> tag.
Value of attribute "moviedb" MUST BE set to the string value of the
PLUGIN_ID (see ex below)
Multiple <id> tags are also supported.
ex:
<movie>
 <id>tt11111</id>
 <id moviedb="allocine">22222</id>
 <id moviedb="filmweb">33333</id>
 <id moviedb="thetvdb">44444</id>
</movie>